# api/swagger.yaml
openapi: 3.0.0
info:
  title: fightpicker API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------------#
# OperationIds conform to the pattern: {version}.{resource}.{http_method}.{action} #
# all lowercase and snake case where required, e.g. v1.users.get.list              #
# ---------------------------------------------------------------------------------#
paths:
...
  /auth/refresh:
    post:
      security: []
      summary: Refresh access token
      operationId: v1.auth.post.refresh
      tags:
        - auth
      description: >
        Issues a new access token using a valid refresh token.
        The refresh token is provided via an HttpOnly cookie.
      responses:
        '200':
          description: Access token refreshed successfully
          headers:
            Set-Cookie:
              description: Rotated refresh token
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Refresh token revoked or account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          format: jwt
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      headers:
        Set-Cookie:
          description: Refresh token cookie
          schema:
            type: string
            example: "refresh_token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Lax; Path=/api/v1/auth; Max-Age=86400"
      required:
        - access_token

    # ----------------------------------#
    #     REUSABLE DATA STRUCTURES      #
    # ----------------------------------#
...
    ErrorEnvelope:
      type: object
      description: Standard API error wrapper
      properties:
        error:
          $ref: '#/components/schemas/ErrorObject'
      required:
        - error
      example:
        error:
          code: "BAD_REQUEST"
          message: "Email is required."
          request_id: "c2f0bad52b9f4fb3"
    ErrorObject:
      type: object
      description: Machine-readable error structure
      properties:
        code:
          type: string
          description: Application-level error code (NOT HTTP status)
          example: "INVALID_INPUT"
        message:
          type: string
          description: Human-readable explanation of the error
          example: "Email is required."
        request_id:
          type: string
          description: Correlation ID for tracing the request
          example: "c2f0bad52b9f4fb3"
      required:
        - code
        - message
        - request_id